Cleaned things up a bit.
